:memo: Add information week 1
Today I wrote my ideas and thoughts for the assignment of this subject. The idea would be a menu for a restaurant purely made in CSS.

My first thought was to make this menu as sleek as possible and build ontop of that with responsiveness. To add a challenge to that I would add at least two colours